working weave subpackage
