+++ src/aiq/observability/register.py
@@ -28,52 +28,6 @@ from aiq.data_models.telemetry_exporter import TelemetryExporterBaseConfig
 logger = logging.getLogger(__name__)
 
 
-def set_wandb_api_key(config_api_key: Optional[str] = None) -> Optional[str]:
-    """
-    Get the W&B API key from various sources in order of priority:
-    1. Config provided key
-    2. WANDB_API_KEY environment variable
-    Returns:
-        The API key if found, None otherwise
-    """
-    if config_api_key:
-        return config_api_key
-    # Check environment variable
-    env_api_key = os.environ.get("WANDB_API_KEY")
-    if env_api_key:
-        return env_api_key
-    return None
-
-
-class WeaveTelemetryExporter(TelemetryExporterBaseConfig, name="weave"):
-    """A telemetry exporter to transmit traces to Weights & Biases Weave using OpenTelemetry."""
-    entity: str = Field(description="The W&B entity/organization.")
-    project: str = Field(description="The W&B project name.")
-    api_key: Optional[str] = Field(
-        default=None,
-        description="Your W&B API key for auth. If not provided, look for WANDB_API_KEY environment variable.")
-
-
-@register_telemetry_exporter(config_type=WeaveTelemetryExporter)
-async def weave_telemetry_exporter(config: WeaveTelemetryExporter, builder: Builder):
-    if config.api_key:
-        set_wandb_api_key(config.api_key)
-
-    import weave
-    _ = weave.init(project_name=f"{config.entity}/{config.project}")
-
-    class NoOpSpanExporter:
-
-        def export(self, spans):
-            return None
-
-        def shutdown(self):
-            return None
-
-    # just yielding None errors with 'NoneType' object has no attribute 'export'
-    yield NoOpSpanExporter()
-
-
 class PhoenixTelemetryExporter(TelemetryExporterBaseConfig, name="phoenix"):
     """A telemetry exporter to transmit traces to externally hosted phoenix service."""
